+++ packages/core/src/blocks/ListItem/CheckListItem/block.ts
@@ -90,7 +90,10 @@ export const createCheckListItemBlockSpec = createBlockSpec(
       // schema.
       const paragraph = document.createElement("p");
 
-      dom.appendChild(checkbox);
+      const div = document.createElement("div");
+      div.contentEditable = "false";
+      div.appendChild(checkbox);
+      dom.appendChild(div);
       dom.appendChild(paragraph);
 
       return {
@@ -148,19 +151,18 @@ export const createCheckListItemBlockSpec = createBlockSpec(
       },
       inputRules: [
         {
-          find: new RegExp(`^\\[\\s*\\]\\s$`),
+          find: /^\s?\[\s*\]\s$/,
           replace() {
             return {
               type: "checkListItem",
               props: {
                 checked: false,
               },
-              content: [],
             };
           },
         },
         {
-          find: new RegExp(`^\\[[Xx]\\]\\s$`),
+          find: /^\s?\[[Xx]\]\s$/,
           replace() {
             return {
               type: "checkListItem",